+++ Samples/Legacy/PWAppDelegate.h
@@ -1,0 +1,14 @@
+//
+//  PWAppDelegate.h
+//  PWMapKitSample
+//
+//  Copyright (c) 2014 Phunware, Inc. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface PWAppDelegate : UIResponder <UIApplicationDelegate>
+
+@property (strong, nonatomic) UIWindow *window;
+
+@end