+++ Samples/Legacy/main.m
@@ -1,0 +1,17 @@
+//
+//  main.m
+//  PWMapKitSample
+//
+//  Copyright (c) 2014 Phunware, Inc. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+#import "PWAppDelegate.h"
+
+int main(int argc, char * argv[])
+{
+    @autoreleasepool {
+        return UIApplicationMain(argc, argv, nil, NSStringFromClass([PWAppDelegate class]));
+    }
+}