+++ Samples/POI Management/POI Management/main.m
@@ -1,0 +1,16 @@
+//
+//  main.m
+//  POI Management
+//
+//  Created by Illya Busigin on 4/13/15.
+//  Copyright (c) 2015 Phunware, Inc. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+#import "AppDelegate.h"
+
+int main(int argc, char * argv[]) {
+    @autoreleasepool {
+        return UIApplicationMain(argc, argv, nil, NSStringFromClass([AppDelegate class]));
+    }
+}